GitHub.App.Token.Refresh
Installation tokens are good for one hour, and this is not configurable.
If you use this token in a long operation (such as iterating over a
large list of things and making API calls along the way), the token will
expire and you'll start getting 401 responses.

```hs
token <- generateInstallationToken creds installationId

for_ repos $ \repo -> do
  makeRequest token repo
```

Generating a token for each item of the list would be one way to avoid
this,

```hs
for_ repos $ \repo -> do
  token <- generateInstallationToken creds installationId
  makeRequest token repo
```

Bug we'd be making a lot of tokens that are unnecessary most of the time.

The new `Refresh` module implements a check-refresh loop, storing a
token in a reference and then giving back the reference to use in each
iteration. Pulling a token from the reference will then always give you an
un-expired token, and the minimum number of tokens necessary will be
generated.

The interface is meant to look much like the non-refreshed example:

```diff
-token <- generateInstallationToken creds installationId
+ref <- refreshing $ generateInstallationToken creds installationId

 for_ repos $ \repo -> do
+  token <- getRefresh ref
   makeRequest token repo
```