generateInstallationTokenScoped
Introduces a new function for specifying repositories or permissions for
the created token. This is accomplished by passing a value of type
`CreateAccessToken` which matches the `POST` body.

```hs
generateInstallationTokenScoped CreateAccessToken
  { repositories = ["foo/bar", ...]
  , repository_ids = [1, ...]
  , permissions = checks Read <> actions Write <> contents Read
  }
```

The type is a `Semigroup` and `Monoid` for better ergonomics of
construction and combination. The `Permissions` type is slightly
complicated, and will be annoying to maintain, but that's so it's
impossible to misuse.